requirements.txt
openai-whisper 
gTTS 
transformers 
sentencepiece 
torch